Fixed Zooshi Windows build for VS2012.
Visual Studio doesn't support brace initialization for complex types. It
also has issues with implicit move constructors and other small nitpicky
syntax things.

Also disabled some warnings that cause the VS build to fail; these will
eventually be fixed in the upstream version of Bullet.

+++ src/states/world_editor_state.cpp
@@ -30,7 +30,8 @@ using mathfu::quat;
 namespace fpl {
 namespace fpl_project {
 
-static const float kEditorViewportAngle = M_PI / 3.0;  // 60 degrees
+static const float kEditorViewportAngle =
+    static_cast<float>(M_PI) / 3.0f;  // 60 degrees
 void WorldEditorState::Initialize(Renderer* renderer, InputSystem* input_system,
                                   editor::WorldEditor* world_editor,
                                   World* world) {
